% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{create_layout}
\alias{create_layout}
\title{create_layout()}
\usage{
create_layout(network)
}
\arguments{
\item{network}{The network w}
}
\value{
Matrix array that gives the coordinates of each node. This should then be parsed as the layout_coords argument in create_network_plots().
}
\description{
Creates a network layout using the Fruchterman–Reingold force-directed layout algorithm, resulting in a balanced, spaced layout.
}
\examples{
m <- matrix(sample(0:1, 25, replace=TRUE), nrow=5)
m[lower.tri(m)] <- t(m)[lower.tri(m)]
diag(m) <- 0
create_layout(m)
}
