% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{create_network_plots}
\alias{create_network_plots}
\title{Title}
\usage{
create_network_plots(
  data_list,
  layout_coords,
  color = "red",
  edge_color = "purple",
  title = "Network Plots"
)
}
\arguments{
\item{title}{}
}
\description{
Title
}
