% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{DKL}
\alias{DKL}
\title{DKL()}
\usage{
DKL(Om.est, Om.true)
}
\arguments{
\item{Om.est}{Estimated precision/adjacency matrix}

\item{Om.true}{True precision/adjacency matrix}
}
\value{
Kullbuck-Liebler divergence between the estimated and true precision matrices.
}
\description{
Calculates the Kullbuck-Liebler divergence between the estimated and true precision matrices.
}
