% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NExON_func.R
\name{NExON}
\alias{NExON}
\title{Title}
\usage{
NExON(
  Ys,
  ts = NULL,
  list_hyper = NULL,
  list_init = NULL,
  tol = 0.001,
  maxit = 1e+05,
  verbose = F,
  debug = F,
  v0_list = NULL
)
}
\arguments{
\item{v0_list}{}
}
\description{
Title
}
