% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{precision}
\alias{precision}
\title{precision()}
\usage{
precision(conf.mat)
}
\arguments{
\item{conf.mat}{A 2x2 confusion matrix}
}
\value{
The precision corresponding to the confusion matrix.
}
\description{
Makes a calculation for precision based on the confusion matrix entries (TP/(TP+FP))
}
\examples{
.
}
