% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_methods.R
\name{select_optimal_v0_t}
\alias{select_optimal_v0_t}
\title{Title}
\usage{
select_optimal_v0_t(
  Y,
  v0_range = 10^seq(log10(0.02), log10(0.2), length.out = 30),
  plot_ = TRUE,
  gamma = 0.15,
  save_outs = F,
  t_range = seq(0, 1, length.out = 1000)
)
}
\arguments{
\item{t_range}{}
}
\description{
Title
}
