% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{sparsity}
\alias{sparsity}
\title{sparsity()}
\usage{
sparsity(network)
}
\arguments{
\item{network}{A precision/adjacency matrix}
}
\value{
Sparsity calculation of the inputted matrix
}
\description{
Calculates the sparsity of a matrix
}
