% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSL_func.R
\name{SSL}
\alias{SSL}
\title{Title}
\usage{
SSL(
  Y,
  list_hyper,
  tol = 0.001,
  maxit = 1e+05,
  verbose = T,
  anneal = NULL,
  track_q = F,
  track_vb_elbo = F,
  set_v0 = 0.05
)
}
\arguments{
\item{set_v0}{}
}
\description{
Title
}
